@@ -10,8 +10,15 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
+    "@react-navigation/stack": "^6.3.17",
     "react": "18.2.0",
-    "react-native": "0.72.3"
+    "react-native": "0.72.3",
+    "react-native-paper": "^5.10.0",
+    "react-native-safe-area-context": "^4.7.1",
+    "react-native-screens": "^3.24.0",
+    "react-native-vector-icons": "^10.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -19,12 +26,14 @@
     "@babel/runtime": "^7.20.0",
     "@react-native/eslint-config": "^0.72.2",
     "@react-native/metro-config": "^0.72.9",
-    "@tsconfig/react-native": "^3.0.0",
+    "@tsconfig/react-native": "^3.0.2",
     "@types/react": "^18.2.18",
     "@types/react-native": "^0.72.2",
     "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
     "babel-jest": "^29.2.1",
-    "eslint": "^8.19.0",
+    "eslint": "^8.46.0",
     "jest": "^29.2.1",
     "metro-react-native-babel-preset": "0.76.7",
     "prettier": "^2.4.1",
