@@ -1,3 +1,6 @@
-module.exports = {
-  presets: ['module:metro-react-native-babel-preset'],
+export const presets = ['module:metro-react-native-babel-preset'];
+export const env = {
+  production: {
+    plugins: ['react-native-paper/babel'],
+  },
 };
